requirements.txt
# GEO Audit System - Core Requirements
# Tested compatible versions

# AI API clients
openai>=1.70.0
anthropic>=0.50.0
google-generativeai>=0.7.0

# HTTP and async
aiohttp>=3.9.0
httpx>=0.27.0
requests>=2.32.0
tenacity>=8.2.0
asyncio-throttle>=1.0.2

# Data processing
pandas>=2.1.0
numpy>=1.24.0
pydantic>=2.6.0
python-dateutil>=2.8.0

# Configuration
pyyaml>=6.0.0
python-dotenv>=1.0.0
urllib3>=1.26.0

# Web scraping
beautifulsoup4>=4.12.0
lxml>=4.9.0
trafilatura>=1.6.0
markdownify>=0.11.0

# Text analysis
textstat>=0.7.0
nltk>=3.8.0

# Visualization and reports
matplotlib>=3.5.0
reportlab>=3.6.0

# Web automation (optional)
selenium>=4.15.0
webdriver-manager>=4.0.0

# Development tools (optional)
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0
